chg: slab component can be used together with disk component
If ngpop = 2 and idisk, ienv and islab are True, then first dust
species is the disk, the second dust species density is equal to
the sum of envelope and slab densities. (The envelope grain size
and opacity is used for the slab too.)

If ngpop = 3 and idisk, ienv and islab are True, then first dust
species is the disk, the second dust species is the envelope and
the third dust species is the slab. The dust opacity model for
the components can be set individually. In this case do not
forget to set three components for gsmin and gsmax!