new: support for upsizing raytraced image.
New impar parameter 'sizeau_ext' added. When this is set and larger
than 'sizeau', then a sizau * sizeau size image is raytraced, then
a zero filled array with the sizeau_ext * sizeau_ext is created, with
its centre filled with the content of the original, smaller scale image.
Both images have to have even number of pixels (and nx = ny must be true).

This procedure helps in speeding up computation when large scales contribute
little to the overall emission.